Facing reality with touch events, redoing logic on gesturestart
When touching with two fingers, iOS triggers single touch event with two
changed touches in it. This requires logic overhaul - scan the changed
touches and instantiating touch for each.

Gesturestart event turned out to be a special case, as you have to
trigger it *before* the second touch is moved. Example: swiping with one
finger and touching with other should trigger zoom, too. Hence the
gesture start change.

Change Drag.capturing logic to reference of the drag object.