Add a new pane to facilitate usage in kendo-bootstrapper
The new pane shows a "Save" button which calls a function on the parent
frame; using this in the kendo-bootstrapper in order to save the new theme
directly rather than ask the user to copy/paste the CSS and save it
manually.

Should not affect normal operation (it only kicks in in the presence of a
KENDO_THEMEBUILDER_OPTIONS variable in the parent frame).