[angular] Breaking: k-ng-model returns the data item, not valueField
For ComboBox, DropDownList and MultiSelect, `k-ng-model` used to place in
the $scope the field indicated by `dataValueField`, making it effectively
the same as `ng-model`.

Since, as per documentation, `k-ng-model` works with "logical values"
instead of the HTML element's value, it makes sense for it to place data
items in scope.

`dataValueField` remains required in order to setup bi-directional binding,
as well as to properly set the HTML element's value.

To get previous behavior, just use `ng-model`.

Close #261