[PDF] Fix poorly rendered borders.
What actually fixes it is pretty simple: use getBoundingClientRect() instead
of getClientRects(), as it returns correct values in
Chrome (getClientRects() has rounded integers).  It took all this patch to
get to this conclusion, though. :( Good thing is that we now optimize some
simple border cases.

Also, better rendering for table cells with collapsed borders in
Chrome (still not perfect though).