[PDF] Fix caching images in Firefox
Close #3899

It appears that the real issue in this case was that FF will trigger onerror
if we set crossOrigin="Anonymous" on a image specified with a data:... url.
Also, because of the hack that reset the src (now removed), an onload was
triggered as well for the same image; counting two images, generation
proceeded, although the sprite image didn't yet manage to load.  Tricky...