Fix tooltip popup still shows when destroyed
Closes #1054:

How to reproduce:

1. Make a tooltip with a delay.
2. Move your mouse on it.
3. Somehow call tooltip.destroy() before the popup shows up.
4. Note that the popup still shows up.

My context is that, the button that shows up the tooltip, when clicked
on will also navigate to another state(like another page, but only part
of page content is changed) in my angular SPA. The popup will still show
up after the click. The popup delay is 1000ms.