Sample code not working for 'depth' option
In the documentation is stated that:
"Note the option will not be applied if "start" option is "lower" than "depth". Always set both and "start" and "depth" options.

But the given example is without the 'start' option, so it is not working correctly. I put both "start" and "depth" set as "year" and now the example is working as expected.