Get rid of Grunt but offer backward compatibility
According to @underlog (see #1390 for more details) Kendo UI Core
migrated from Grunt to Gulp, but some leftovers are still present.

This is my attempt to clean a bit the Grunt file definition, this change
is backward compatibility, this means that if someone execute the
`grunt jshint` command will work, but under the hood, we use Gulp tasks.