Force re-render of select <options> tag if one of the selected dataItems is not present in the current data view. Fix telerik/kendo-ui-core#1618
The issue occurs because at present the MultiSelect widget does not
re-render its <options> tags when the List fires its change event. This
works fine with the StaticList, however the logic breaks with the
VirtualList where the developer can select a value that is not part of
the current data page. In that scenario the VirtualList makes data
prefetch to obtain the corresponding dataItem but does not change the
current data page, hence the widget will not update its <options>.

As a result when the VirtualList fires its change event and the MultiSelect
calls `_selectValue` and the `_customOptions` array does not include the
custom value.

Later when the tag is about to be removed the widget cannot resolve it because and
threats it as a new tag that has to be added. This is the reason for the
duplication.

Current solution checks if there is already generated <option> for the
added items at the List change event handler. If there isn't one the
MultiSelect will re-render its <options> list.

The ideal solution would be to change the way <option> list is created.
Instead of rendering an option tag for each element from the data view
we can render <option> elements only for the selected dataItems.