Ensure that the _ranges are cleared. Fixes telerik/kendo-ui-core#1826
The change is required, because the datasource cannot recognize the subsequent requests as standalone, which results in mutiple populated ranges. The fix basically ensures that the _ranges are indeed cleared.
Unfortunately, the change cannot be tested as it is reproducible only with moderate filtering frequence. If you type fast or slow, the issue is not visible