Unbind from change events when removing item
When a Model is being added, a bind to the "change" event from that Model is being set, but when this exactly same item is removed, the unbind it is not being call. This is triggering multiple "change" events when adding and removing the same Model from a collection several times.