feat(masked): Use input event for user content
Use exclusively the input event for tracking user content and applying the mask. Track the changed content and update it according to the mask. This makes the implementation of the MaskTextBox masking more consistent and improves masking on mobile devices.

The input event implementation of the MaskedTextBox should also fix the following issue: telerik/kendo#6771

IE[9,10] do not support the input event, so simulate it through tracking other events:

* keydown - for general input through the keyboard
* mouseup - for paste from context menu
* paste - for paste over keyboard shortcuts or browser's menu
* drop - when dropping new content

In Chrome on Android the key* DOM events do not expose which key was pressed - https://bugs.chromium.org/p/chromium/issues/detail?id=118639. It seems that this will not be fixed. This feature is needed to track if backspace has been used to delete content. Workaround through an heuristics on the content changes was introduced.

In WebKit browsers the input event is triggered twice when drag&dropping from the same input. Therefore we should skip changing the input if content gets removed and dropping at the same time.

Prevent the ng-model binding from happening during the input event. The approach is to detach the "input" event altogether, which is the event the ngjs engine uses to track changes on the input value. In case this is not acceptable behavior, consider the option of manually triggering the "right" event to update the ng-model.

Chrome raises the change event even though the value is not different between focused and blurred. It seems that the Chrome browser flags an input as changed when there is user input although it may have been the same after explicitly removing the new user input.

Caret positioning on Windows Phone and Android devices needs timeout.

teleri/kendo#6739