Add sample code to work around grouping in the original dataSource
The sample throws an error if the original grid data has a grouping applied. The view() data returned in line 70 returns the original data sorted with the new sort configuration but also grouped by the original grouping.  Setting the checkSource group configuration to null fixes this case. This does not affect any group configuration in the original dataSource.