Update datasource.md
The text providing possible values for the type in the list of fields in Schema.Model.Fields should be lowercase.  Having uppercase values implies that uppercase values will work for the type, and that is not true in every case.  Specifically, uppercase values in the type field will cause the row filtering to not properly apply the autocomplete widget to inline row editing text boxes in the filter row.